Update to fedora 38 and mkosi v22
This build the same IPA code just with a newer fedora image and mkosi
version.

Fedora 38 is the latest version with python 3.11 as the default.
Newer versions run with python 3.12, which is not supported by
ironic, much less our older version of it.

Installing python 3.11 on later version adds to the image size.