Script to create an ESP image
Ironic needs some ESP image for setting up a bootable iso.
This creates it from scratch for the OS it runs on.

Keep in mind, that the choice of the OS has deployment
implications, as the deployment needs to know the OS
as the path of the grub configuration depends on that.